Update to fix import error
Due to a testing error in depoyment on tsuru, namely that when you push to a Gandalf git repo without
specificly stating which local branch you want pushed, you push master, a cyclical import error crept in.

This commit reverts the import issue (caused by a cleanup change) and fixes the use of environment variables
for the metrics service. Mitigation of issues like this in the future would be to have a lightweight set
of local tests.