Add tsuru healthcheck config
Which will ensure that `GET /healthcheck` returns a 200 response before new
units are are added to the router. This means that application deploys and
restarts have zero downtime.

As noted in the preceding commit, the app will fail to start if it doesn't
have the environment variables or can't connect to PostgreSQL, so we need to
ensure those things are configured correctly for the demo otherwise the
deploy will also fail.