Add /healthcheck endpoint
Which will return an HTTP 200 status if the app is running. This is very
simple, it intentionally does not confirm that the app can use PostgreSQL,
because that shouldn't be a dependency of starting or deploying.

HOWEVER: the app does currently fail to start if it doesn't have the
environment variables from Tsuru or it's unable to connect to PostgreSQL
when it first starts.