Set defaults for pg_* environment variables
In order to allow the app to start and serve `/healthcheck` when the
environment variables have not been set. Otherwise it constantly fails
Tsuru's healthchecks when the app is:

- started (deployed) before `tsuru service-bind`
- restarted (automatically) after `tsuru service-unbind`

The latter has been extending the time our smoke tests take to run from
approx 50 seconds to 3 minutes because it keeps waiting for the healthcheck
to pass.

We should be able to start the app and bring it into service even if the
database connection isn't working.