Allow table creation to fail on start
In order to allow the app to start and serve `/healthcheck` when the
database is down or the correct environment variables have not been
configured. We shouldn't prevent a deployment or restart in these
circumstances, even though the app may not function completely.

In such a case an exception like the following will be logged to STDOUT and
picked up by Tsuru:

    2015-07-14 13:58:40 +0100 [web][e17e40337880]: (OperationalError) could not connect to server: Connection refused
    2015-07-14 13:58:40 +0100 [web][e17e40337880]:  Is the server running on host "localhost" (::1) and accepting
    2015-07-14 13:58:40 +0100 [web][e17e40337880]:  TCP/IP connections on port 5432?
    2015-07-14 13:58:40 +0100 [web][e17e40337880]: could not connect to server: Connection refused
    2015-07-14 13:58:40 +0100 [web][e17e40337880]:  Is the server running on host "localhost" (127.0.0.1) and accepting
    2015-07-14 13:58:40 +0100 [web][e17e40337880]:  TCP/IP connections on port 5432?
    2015-07-14 13:58:40 +0100 [web][e17e40337880]:  None None

If this is the first time that the application has been deployed and the
tables haven't been created before then it will require a restart, which in
the case of missing/incorrect environments will happen after `tsuru
service-bind`, in order to run `init_db()` again. This should be apparent
from the error pages produced the app though.

This, combined with the previous commit, will reduce the time that the smoke
tests take to run.