Reinstate init_db call broken in move to gunicorn
In commit 8347bef0b8966aa461714c97f069838ae08a1b01, the move to gunicorn missed the fact that we
registered a database schema creation hook. Thus, when testing against an already existing application
the change worked perfectly. However, when the smoke tests in CI started a new app instance, the build
broke.

This commit moves the hook into the app declaration and now enables the db_init command to be called
when using gunicorn.